Refactor duplicated online retriever code
